fixed stupid error in rapidxml::parse_error
the class has data pointers and did not have a copy constructor and did
not have an assignment operator. thus the compil;er gave a warning, but
sionce the class name contains error, it trats the warning as an error.
So i implemented the missing functions. and by saving the file a bunch
of the whitespace got changes.