thread pool can now reduce active threads.
not perfect, may take a while for certain threads to wait for condition.
if there are 5 threads, and only 2 are allowed to to work the last 3
threads may continue to work on long tasks, while the first 2 (which
logically should not be executing because of the activity limit) may
work through multiple small tasks.

this is because they mechanism by which the execution is witheld is by
thread ID, so threads ID'd as 0 & 1, are less than 2 (the limit) where
threads ID'd as 2, 3, & 4 are not and tus wait to be allowed to work,
(this uses the above example as well)